@@ -1,12 +1,12 @@
 import { Injectable } from '@angular/core';
-import { Http, Headers, RequestOptions } from '@angular/http';
+import { HttpClient, HttpHeaders,  } from '@angular/common/http';
 
 @Injectable()
 export class UserService {
-    private headers = new Headers({ 'Content-Type': 'application/json', 'charset': 'UTF-8' });
-    private options = new RequestOptions({ headers: this.headers });
+    private headers = new HttpHeaders({ 'Content-Type': 'application/json', 'charset': 'UTF-8' });
+    private options = { headers: this.headers };
 
-    constructor(private http: Http) { }
+    constructor(private http: HttpClient) { }
 
     login(credentials) {
         return this.http.post('/api/login', JSON.stringify(credentials), this.options);
