@@ -1,18 +1,21 @@
 import { Injectable } from '@angular/core';
-import { Http, Headers, RequestOptions } from '@angular/http';
+import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { BehaviorSubject } from 'rxjs/BehaviorSubject';
 
 import { AuthService } from '../../login/services/auth.service';
 
 @Injectable()
 export class DashboardData {
-    private headers = new Headers({ 'Content-Type': 'application/graphql', 'charset': 'UTF-8' });
-    private options = new RequestOptions({ headers: this.headers });
+    private headers = new HttpHeaders({
+        'Content-Type': 'application/graphql',
+        'charset': 'UTF-8'
+    });
+    private options = { headers: this.headers };
     dataChange = new BehaviorSubject([]);
 
     get data() { return this.dataChange.value; }
 
-    constructor(private http: Http, private authService: AuthService) {
+    constructor(private http: HttpClient, private authService: AuthService) {
         const user = authService.getUser();
 
         const query = `{
@@ -26,7 +29,8 @@ export class DashboardData {
 
         this.http
             .post('/api/graphQL', query, this.options)
-            .map(res => res.json().data.orders)
-            .forEach(item => this.dataChange.next(item));
+            .subscribe(res => {
+                this.dataChange.next(res['data'].orders);
+            });
     }
 }
